[FIX] BKHV-56

+++ src/app/apps/shopping-cart/shopping-cart-details/shopping-cart-details.component.ts
@@ -37,12 +37,13 @@ export class ShoppingCartDetailsComponent {
   }
 
   removeBook(bookId: string) {
-    this._shoppingCartService.removeBookFromShoppingCart(bookId);
-    this.isLoading = true;
-    this._notifierService.notify('success', 'Book removed successfully');
-    console.log(this.shoppingCartInfo);
-    this.getShoppingCartInfo();
-    location.reload();
+    if (confirm('Are you sure you want to remove the item from your cart?')) {
+      this._shoppingCartService.removeBookFromShoppingCart(bookId);
+      this.isLoading = true;
+      this._notifierService.notify('success', 'Book removed successfully');
+      this.getShoppingCartInfo();
+      location.reload();
+    }
   }
 
   checkout() {
